requirements.txt
alembic==0.8.6
autopep8==1.2.4
blinker==1.4
coverage==4.1
docopt==0.6.2
enum==0.4.6
enum34==1.1.6
Flask==0.10.1
Flask-Assets==0.11
Flask-Cache==0.13.1
Flask-Limiter==0.9.3
Flask-Migrate==1.8.0
Flask-MySQL==1.3
Flask-OpenID==1.2.5
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Testing==0.4.2
itsdangerous==0.24
Jinja2==2.8
limits==1.1.1
Mako==1.0.4
MarkupSafe==0.23
monotonic==1.1
MySQL-python==1.2.5
pep8==1.7.0
pyflakes==1.2.3
python-editor==1.0
python-openid==2.2.5
python-valve==0.1.1
requests==2.10.0
six==1.10.0
SQLAlchemy==1.0.13
webassets==0.11.1
Werkzeug==0.11.9
WTForms==2.1
